requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.11.0

# Vector search and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# LLM backends
openai>=1.0.0
anthropic>=0.18.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Cryptography
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: LangChain for RAG
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0

# Development
black>=23.0.0
mypy>=1.5.0
ruff>=0.1.0
